+++ flow-typed/platform.js
@@ -37,6 +37,7 @@ declare type AppContext = {
 				org?: 1,
 		  |}
 		| {| id: 0 |},
+	geo: GeoLocation,
 	initialNow: number, // timestamp that React-Intl uses to initialize date strings on initial client render
 	isQL: boolean,
 	isProdApi: boolean, // whether the data is coming from Prod instead of Dev DB
@@ -149,6 +150,8 @@ declare type PlatformRoute = AsyncPlatformRoute | StaticPlatformRoute;
 declare type GeoLocation = {
 	country?: string,
 	region?: string,
+	city?: string,
+	latlon?: Array<number>,
 };
 
 // See https://docs.launchdarkly.com/docs/node-sdk-reference#section-users